# docs/contracts/openapi-area-jugador.yaml
openapi: 3.1.0
info:
  title: ReservaYa API — Contrato de Área Jugador y Vitrina Pública
  description: >
    Especificación OpenAPI de los contratos compartidos entre la landing pública
    (apps/web), el panel de usuario (apps/web) y
    el backend (.NET 10 ReservaFacil.Api) para el área del jugador bajo
    arquitectura desacoplada.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Backend local .NET
  - url: http://localhost:3000
    description: Proxy Next.js local
  - url: https://reservaya-api.onrender.com
    description: Backend Render producción
paths:
...
  /api/resenas/publicas:
    get:
      summary: Reseñas públicas y valoración agregada por complejo
      parameters:
        - name: complejoId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Promedio de valoración y lista de opiniones
          content:
            application/json:
              schema:
                type: object
                required: [promedio, total, resenas]
                properties:
                  promedio:
                    type: number
                    example: 4.8
                  total:
                    type: integer
                    example: 24
                  resenas:
                    type: array
                    items:
                      $ref: '#/components/schemas/Resena'
...
  /api/resenas:
    post:
      summary: Calificar un complejo tras completar una reserva
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [complejoId, puntuacion]
              properties:
                complejoId:
                  type: string
                puntuacion:
                  type: integer
                  minimum: 1
                  maximum: 5
                comentario:
                  type: string
      responses:
        '200':
          description: Reseña guardada
          content:
            application/json:
              schema:
                type: object
                required: [ok, resena]
                properties:
                  ok:
                    type: boolean
                  resena:
                    $ref: '#/components/schemas/Resena'
...
components:
  schemas:
...
    Resena:
      type: object
      required: [puntuacion]
      properties:
        id:
          type: string
        puntuacion:
          type: integer
        comentario:
          type: string
          nullable: true
        respuestaDueno:
          type: string
          nullable: true
        creadoEn:
          type: string
        usuario:
          type: object
          properties:
            nombre:
              type: string
...